Add openall and open_by_url methods
